--- src/components/cardComponent/CardDetail.scss
@@ -1,100 +1,0 @@
-@import '@/styles/ColorList.scss';
-
-.card {
-  border: 1px;
-  background: white;
-  border: 1px solid Grayscale-200;
-
-  &[data-size='S'] {
-    width: 343px;
-    height: 190px;
-    border-radius: 12px;
-  }
-  &[data-size='Group63'] {
-    width: 704px;
-    height: 260px;
-    border-radius: 16px;
-  }
-  &[data-size='L'] {
-    width: 1140px;
-    height: 260px;
-    border-radius: 16px;
-  }
-}
-.cardImage {
-  box-shadow: 0px 2px 20px 0px #0000000a;
-  &[data-size='S'] {
-    width: 58px;
-    height: 209px;
-    left: 20px;
-  }
-  &[data-size='Group63'] {
-    width: 84px;
-    height: 302px;
-    left: 60px;
-  }
-  &[data-size='L'] {
-    width: 58px;
-    height: 209px;
-    //top: 93px;
-    left: 100px;
-  }
-}
-.wine {
-  width: 300px;
-  height: 111px;
-  //top: 94px;
-  //left: 244px;
-  gap: 20px;
-}
-.wineName {
-  &[data-size='S'] {
-    width: 200px;
-    height: 48px;
-    //top: 52px;
-    //left: 98px;
-    font-family: Pretendard;
-    font-size: 20px;
-    font-weight: 600;
-    line-height: 23.87px;
-    text-align: left;
-    background: Grayscale-400;
-  }
-  /*&[data-size="Group63"]{
-    width: 
-  }*/
-  &[data-size='L'] {
-    width: 300px;
-    height: 72px;
-    font-family: Pretendard;
-    font-size: 30px;
-    font-weight: 600;
-    line-height: 35.8px;
-    text-align: left;
-    background: Grayscale-400;
-  }
-}
-/*
-.wineDesc{
-  &[data-size="S"]{
-    width: 
-  }
-  &[data-size="Group63"]{
-    width: 
-  }
-  &[data-size="L"]{
-    width: 
-  }
-}
-.winePrice{
-  &[data-size="S"]{
-    width: 
-  }
-  &[data-size="Group63"]{
-    width: 
-  }
-  &[data-size="L"]{
-    width: 
-  }
-}
-*/
