+++ src/components/cardmylist/card.scss
@@ -1,0 +1,3 @@
+@import '@/styles/ColorList.scss';
+@import '@/styles/Font-Mixins.scss';
+@import '../../app/globals.scss';
