+++ src/components/inputComponent/Input.tsx
@@ -1,27 +1,19 @@
 'use client';
 
-import React, { useState } from 'react';
+import React from 'react';
 import './Input.scss';
+import { InputProps } from '@/types/Input';
 
-interface Props {
-  type: 'email' | 'password' | 'text';
-  size: 'S' | 'L';
-  placeholder?: string;
-  inputname: string;
-}
-
-const Input: React.FC<Props> = ({ type, size, placeholder = '입력하시오.', inputname = '입력하시오.' }) => {
-  const [inputvalue, setInputvalue] = useState('');
-
+const Input: React.FC<InputProps> = ({ size, placeholder = '입력하시오.', inputname = '입력하시오.', defaultValue = '', ...rest }) => {
   return (
-    <form className="form" data-size={size} data-type={type}>
-      {inputname != '' && (
-        <label className="inputname" data-size={size}>
+    <div className="form" data-size={size}>
+      {inputname && (
+        <p className="inputname" data-size={size}>
           {inputname}
-        </label>
+        </p>
       )}
-      <input className="input" data-size={size} type={type} placeholder={placeholder} value={inputvalue} onChange={(e) => setInputvalue(e.target.value)} />
-    </form>
+      <input className="input" data-size={size} placeholder={placeholder} defaultValue={defaultValue} {...rest} />
+    </div>
   );
 };
