+++ src/types/Input.tsx
@@ -1,0 +1,7 @@
+import { HtmlHTMLAttributes, InputHTMLAttributes } from 'react';
+
+export interface InputProps extends Omit<InputHTMLAttributes<HTMLInputElement>, 'size'> {
+  inputname: string;
+  size: 'S' | 'L';
+  defaultvalue?: string;
+}
